Prepare pages for multiple examples
