+++ playground/app.vue
@@ -1,63 +1,3 @@
-<script setup lang="ts">
-import type { FetchError } from 'ofetch'
-import type { JsonPlaceholderComment } from './types'
-
-const route = useRoute()
-
-// Intended for similar use cases as `useFetch`
-const { data, pending, error } = await useJsonPlaceholderData<JsonPlaceholderComment>(
-  'comments',
-  {
-    query: computed(() => ({
-      postId: `${route.query.postId || 1}`,
-    })),
-    onResponse({ response }) {
-      if (process.server)
-        return
-      // eslint-disable-next-line no-console
-      console.log(response._data)
-    },
-  },
-)
-
-// eslint-disable-next-line no-console
-watch(error, value => console.log(value))
-
-async function incrementPostId() {
-  await navigateTo({
-    query: {
-      postId: `${Number(route.query.postId || 1) + 1}`,
-    },
-  })
-  // eslint-disable-next-line no-console
-  console.log('Post ID:', route.query.postId)
-}
-
-const formResponse = ref()
-
-// Intended for similar use cases as `$fetch`
-async function onSubmit() {
-  try {
-    formResponse.value = await $jsonPlaceholder('posts', {
-      method: 'POST',
-      body: {
-        title: 'foo',
-        body: 'bar',
-        userId: 1,
-      },
-    })
-
-    // eslint-disable-next-line no-console
-    console.log('formResponse:', formResponse.value)
-  }
-  catch (e) {
-    console.error('statusCode:', (e as FetchError).statusCode)
-    console.error('statusMessage:', (e as FetchError).statusMessage)
-    console.error('data:', (e as FetchError).data)
-  }
-}
-</script>
-
 <template>
   <Head>
     <Title>nuxt-api-party</Title>
@@ -65,37 +5,16 @@ async function onSubmit() {
   </Head>
 
   <header>
-    <h1>nuxt-api-party</h1>
+    <NuxtLink to="/">
+      <h1>nuxt-api-party</h1>
+    </NuxtLink>
     <p>
       Requests are proxied by a Nuxt server route and passed back to the client. The playground uses <a href="https://jsonplaceholder.typicode.com/">{JSON} Placeholder</a>
-      as an example API. The dynamic composables <code>$jsonPlaceholder</code> and <code>useJsonPlaceholderData</code> are generated by the module.
+      and <a href="https://petstore3.swagger.io/">Swagger Petstore</a> as example APIs. The dynamic composables <code>$jsonPlaceholder</code> and <code>useJsonPlaceholderData</code> are generated by the module.
     </p>
   </header>
 
   <main>
-    <h2>$jsonPlaceholder</h2>
-    <p>Responses are <strong>not</strong> cached by default.</p>
-    <blockquote>(Imagine form fields here)</blockquote>
-    <p>
-      <button @click="onSubmit()">
-        Submit
-      </button>
-    </p>
-    <pre v-if="formResponse">{{ JSON.stringify(formResponse, undefined, 2) }}</pre>
-    <hr>
-
-    <h2>useJsonPlaceholderData</h2>
-    <p>Responses are cached by default.</p>
-    <p>
-      Status:
-      <mark v-if="pending">pending</mark>
-      <code v-else>fetched</code>
-    </p>
-    <p>
-      <button @click="incrementPostId()">
-        Increment Post ID
-      </button>
-    </p>
-    <pre>{{ JSON.stringify(data, undefined, 2) }}</pre>
+    <NuxtPage />
   </main>
 </template>
