+++ playground/pages/index.vue
@@ -1,0 +1,10 @@
+<template>
+  <h2>Playground</h2>
+  <ul>
+    <li>
+      <NuxtLink to="/jsonPlaceholder">
+        {JSON} Placeholder
+      </NuxtLink>
+    </li>
+  </ul>
+</template>
